t/xml/entities-2.dat

#data
<!DOCTYPE a[
<!ENTITY%>
]><a/>
#errors
2;9;no space before md name;;;m
2;10;no md name;;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities

#data
<!DOCTYPE a[
<!ENTITY%
#errors
2;9;no space before md name;;;m
2;9;unclosed md;;;m
2;9;unclosed internal subset;;;m
2;9;no root element;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY %>
]><a/>
#errors
2;11;no md name;;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities

#data
<!DOCTYPE a[
<!ENTITY %
#errors
2;10;unclosed md;;;m
2;10;unclosed internal subset;;;m
2;10;no root element;;;m
#document
| <!DOCTYPE a>
#entities

#data escaped
<!DOCTYPE a[
<!ENTITY % abc  \u0009
#errors
2;17;unclosed md;;;m
2;17;unclosed internal subset;;;m
2;17;no root element;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY % abc  #
#errors
2;17;string after DOCTYPE name;;;m
2;17;unclosed internal subset;;;m
2;17;no root element;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY % abc PUBLIC
#errors
2;21;unclosed md;;;m
2;21;unclosed internal subset;;;m
2;21;no root element;;;m
#document
| <!DOCTYPE a>
#entities

#data escaped
<!DOCTYPE a[
<!ENTITY % abc PUBLIC\u0020
#errors
2;22;unclosed md;;;m
2;22;unclosed internal subset;;;m
2;22;no root element;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY % abc PUBLIC #
#errors
2;23;string after PUBLIC;;;m
2;23;unclosed internal subset;;;m
2;23;no root element;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY % abc SYSTEM
#errors
2;21;unclosed md;;;m
2;21;unclosed internal subset;;;m
2;21;no root element;;;m
#document
| <!DOCTYPE a>
#entities

#data escaped
<!DOCTYPE a[
<!ENTITY % abc SYSTEM\u0020
#errors
2;22;unclosed md;;;m
2;22;unclosed internal subset;;;m
2;22;no root element;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY % abc PUBLIC #
#errors
2;23;string after PUBLIC;;;m
2;23;unclosed internal subset;;;m
2;23;no root element;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY % abc PUBLIC "a"x
#errors
2;26;string after PUBLIC literal;;;m
2;26;unclosed internal subset;;;m
2;26;no root element;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY % abc PUBLIC "a" x
#errors
2;27;string after PUBLIC literal;;;m
2;27;unclosed internal subset;;;m
2;27;no root element;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY % abc SYSTEM "a"x
#errors
2;26;string after SYSTEM literal;;;m
2;26;unclosed internal subset;;;m
2;26;no root element;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY % abc SYSTEM "a" x
#errors
2;27;string after SYSTEM literal;;;m
2;27;unclosed internal subset;;;m
2;27;no root element;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY % abc PUBLIC "a" "b"x
#errors
2;30;string after SYSTEM literal;;;m
2;30;unclosed internal subset;;;m
2;30;no root element;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY % abc PUBLIC "a" "b" x
#errors
2;31;string after SYSTEM literal;;;m
2;31;unclosed internal subset;;;m
2;31;no root element;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY % bc PUBLIC>
]><a/>
#errors
2;21;no PUBLIC literal;;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities
| <!ENTITY % bc "" "" "" >

#data
<!DOCTYPE a[
<!ENTITY % bc PUBLIC >
]><a/>
#errors
2;22;no PUBLIC literal;;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities
| <!ENTITY % bc "" "" "" >

#data
<!DOCTYPE a[
<!ENTITY % bc SYSTEM>
]><a/>
#errors
2;21;no SYSTEM literal;;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities
| <!ENTITY % bc "" "" "" >

#data
<!DOCTYPE a[
<!ENTITY % bc SYSTEM >
]><a/>
#errors
2;22;no SYSTEM literal;;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities
| <!ENTITY % bc "" "" "" >

#data
<!DOCTYPE a[
<!ENTITY % bc PUBLIC "x" "y">
]><a/>
#errors
#document
| <!DOCTYPE a>
| <{} a>
#entities
| <!ENTITY % bc "" "x" "y" >

#data
<!DOCTYPE a[
<!ENTITY % bc PUBLIC "x""y">
]><a/>
#errors
2;25;no space before system literal;;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities
| <!ENTITY % bc "" "x" "y" >

#data
<!DOCTYPE a[
<!ENTITY % bc PUBLIC "x"'y'>
]><a/>
#errors
2;25;no space before system literal;;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities
| <!ENTITY % bc "" "x" "y" >

#data
<!DOCTYPE a[
<!ENTITY % bc PUBLIC 'x'"y">
]><a/>
#errors
2;25;no space before system literal;;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities
| <!ENTITY % bc "" "x" "y" >

#data
<!DOCTYPE a[
<!ENTITY % bc PUBLIC 'x''y'>
]><a/>
#errors
2;25;no space before system literal;;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities
| <!ENTITY % bc "" "x" "y" >

#data
<!DOCTYPE a[
<!ENTITY % bc PUBLIC "x" "y" NDATA#
#errors
2;30;string after SYSTEM literal;;;m
2;35;no root element;;;m
2;35;unclosed internal subset;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY % bc PUBLIC "x" "y" NDATA #
#errors
2;30;string after SYSTEM literal;;;m
2;36;no root element;;;m
2;36;unclosed internal subset;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY % bc PUBLIC "x" "y" NDATA z#
#errors
2;30;string after SYSTEM literal;;;m
2;37;no root element;;;m
2;37;unclosed internal subset;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY % bc PUBLIC "x" "y" NDATA z #
#errors
2;30;string after SYSTEM literal;;;m
2;38;no root element;;;m
2;38;unclosed internal subset;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY % bc PUBLIC "x" "y" NDATA z>
]><a/>
#errors
2;30;string after SYSTEM literal;;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities
| <!ENTITY % bc "" "x" "y" >

#data
<!DOCTYPE a[
<!ENTITY % b ""
#errors
2;15;no root element;;;m
2;15;unclosed internal subset;;;m
2;15;unclosed md;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY % b ''
#errors
2;15;no root element;;;m
2;15;unclosed internal subset;;;m
2;15;unclosed md;;;m
#document
| <!DOCTYPE a>
#entities

#data escaped
<!DOCTYPE a[
<!ENTITY % b ""\u0020
#errors
2;16;no root element;;;m
2;16;unclosed internal subset;;;m
2;16;unclosed md;;;m
#document
| <!DOCTYPE a>
#entities

#data escaped
<!DOCTYPE a[
<!ENTITY % b ''\u0020
#errors
2;16;no root element;;;m
2;16;unclosed internal subset;;;m
2;16;unclosed md;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY % b "">
]><a/>
#errors
#document
| <!DOCTYPE a>
| <{} a>
#entities
<!ENTITY % b "" "" "" >

#data
<!DOCTYPE a[
<!ENTITY % b ''>
]><a/>
#errors
#document
| <!DOCTYPE a>
| <{} a>
#entities
<!ENTITY % b "" "" "" >

#data
<!DOCTYPE a[
<!ENTITY % b "xxxxx">
]><a/>
#errors
#document
| <!DOCTYPE a>
| <{} a>
#entities
<!ENTITY % b "xxxxx" "" "" >

#data
<!DOCTYPE a[
<!ENTITY % b 'xxxxx'>
]><a/>
#errors
#document
| <!DOCTYPE a>
| <{} a>
#entities
<!ENTITY % b "xxxxx" "" "" >

#data
<!DOCTYPE a[
<!ENTITY % b "yyy" >
]><a/>
#errors
#document
| <!DOCTYPE a>
| <{} a>
#entities
<!ENTITY % b "yyy" "" "" >

#data
<!DOCTYPE a[
<!ENTITY % b 'yyy' >
]><a/>
#errors
#document
| <!DOCTYPE a>
| <{} a>
#entities
<!ENTITY % b "yyy" "" "" >

#data
<!DOCTYPE a[
<!ENTITY % b "a&#x24;b">
]><a/>
#errors
#document
| <!DOCTYPE a>
| <{} a>
#entities
<!ENTITY % b "a$b" "" "" >

#data
<!DOCTYPE a[
<!ENTITY % b '&#32;'>
]><a/>
#errors
#document
| <!DOCTYPE a>
| <{} a>
#entities
<!ENTITY % b " " "" "" >

#data
<!DOCTYPE a[
<!ENTITY % b "&amp;abc">
]><a/>
#errors
#document
| <!DOCTYPE a>
| <{} a>
#entities
<!ENTITY % b "&amp;abc" "" "" >

#data
<!DOCTYPE a[
<!ENTITY % b "c">
<!ENTITY % b "d">
]><a/>
#errors
3;1;duplicate para entity decl;;b;w
#document
| <!DOCTYPE a>
| <{} a>
#entities
<!ENTITY % b "c" "" "" >

#data
<!DOCTYPE a[
<!ENTITY % b "c">
<!ENTITY b "d">
]><a/>
#errors
#document
| <!DOCTYPE a>
| <{} a>
#entities
<!ENTITY b "d" "" "" >
<!ENTITY % b "c" "" "" >

#data
<!DOCTYPE a[
<!ENTITY % b SYSTEM "x" >
%c;
<!ENTITY % d SYSTEM "y" >
]><a/>
#errors
3;1;stop processing;;;i
3;2;string in internal subset;;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities
| <!ENTITY % b "" "" "x" >

#data
<?xml version="1.0" standalone="yes"?>
<!DOCTYPE a[
<!ENTITY % b SYSTEM "x" >
%c;
<!ENTITY % d SYSTEM "y" >
]><a/>
#errors
4;2;string in internal subset;;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities
| <!ENTITY % b "" "" "x" >
| <!ENTITY % d "" "" "y" >
